platform_ext: convert from INTERFACE to static library; add Window/WindowApp factory system
- Convert platform_ext from header-only INTERFACE to static library
- Add globals.cpp with Window/WindowApp factory registration and creation
- Add WindowFactory/WindowAppFactory typedefs and Register/Create functions
- Add virtual destructor to WindowApp
- Add new C API module (oge_c_api) with opaque C wrappers for Window/WindowApp
- Update SDL3 backend to use std::string_view and uint64_t params

+++ engine/modules/platform_ext_sdl3/src/sdl3_window.hpp
@@ -15,7 +15,7 @@ using namespace input;
 class SDL3GameWindow : public Window
 {
    public:
-    SDL3GameWindow(std::string name, int width, int height);
+    SDL3GameWindow(std::string name, uint64_t width, uint64_t height);
     ~SDL3GameWindow();
 
     void Run(WindowApp&) override;
